revert: back out save-receipts-on-signout (PR 94973) for blocker 95957
