--- src/libs/fileDownload/hasGalleryWritePermission.android.ts
@@ -1,26 +1,0 @@
-import {PermissionsAndroid, Platform} from 'react-native';
-
-function hasGalleryWritePermission(): Promise<boolean> {
-    // On Android API Level 33 and above, these permissions do nothing and always return 'never_ask_again'
-    // More info here: https://stackoverflow.com/a/74296799
-    if (Number(Platform.Version) >= 33) {
-        return Promise.resolve(true);
-    }
-
-    const writePromise = PermissionsAndroid.check(PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE);
-    const readPromise = PermissionsAndroid.check(PermissionsAndroid.PERMISSIONS.READ_EXTERNAL_STORAGE);
-
-    return Promise.all([writePromise, readPromise]).then(([hasWritePermission, hasReadPermission]) => {
-        if (hasWritePermission && hasReadPermission) {
-            return true;
-        }
-
-        return PermissionsAndroid.requestMultiple([PermissionsAndroid.PERMISSIONS.READ_EXTERNAL_STORAGE, PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE]).then(
-            (status) =>
-                status[PermissionsAndroid.PERMISSIONS.READ_EXTERNAL_STORAGE] === PermissionsAndroid.RESULTS.GRANTED &&
-                status[PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE] === PermissionsAndroid.RESULTS.GRANTED,
-        );
-    });
-}
-
-export default hasGalleryWritePermission;
