--- src/libs/fileDownload/hasGalleryWritePermission.ts
@@ -1,6 +1,0 @@
-/** Type-resolution base; never called on web (no device gallery). */
-function hasGalleryWritePermission(): Promise<boolean> {
-    return Promise.resolve(false);
-}
-
-export default hasGalleryWritePermission;
