--- src/libs/savePendingReceiptsToGallery/index.ts
@@ -1,4 +1,0 @@
-import {getSaveablePendingReceiptRequests} from './getPendingReceiptRequests';
-import saveReceiptsToGallery from './saveReceiptsToGallery';
-
-export {getSaveablePendingReceiptRequests, saveReceiptsToGallery};
