+++ src/pages/settings/InitialSettingsPage.tsx
@@ -35,11 +35,9 @@ import {resetExitSurveyForm} from '@libs/actions/ExitSurvey';
 import {closeReactNativeApp} from '@libs/actions/HybridApp';
 import {hasPartiallySetupBankAccount, hasPersonalBankAccountMissingInfo} from '@libs/BankAccountUtils';
 import {hasPendingExpensifyCardAction, hasVirtualExpensifyCardMissingPersonalDetails} from '@libs/CardUtils';
-import Log from '@libs/Log';
 import createDynamicRoute from '@libs/Navigation/helpers/dynamicRoutesUtils/createDynamicRoute';
 import useIsSidebarRouteActive from '@libs/Navigation/helpers/useIsSidebarRouteActive';
 import Navigation from '@libs/Navigation/Navigation';
-import {getSaveablePendingReceiptRequests, saveReceiptsToGallery} from '@libs/savePendingReceiptsToGallery';
 import {useIsAgentAccount} from '@libs/SessionUtils';
 import {getFreeTrialText, hasSubscriptionRedDotError} from '@libs/SubscriptionUtils';
 import type {SkeletonSpanReasonAttributes} from '@libs/telemetry/useSkeletonSpan';
@@ -112,11 +110,7 @@ type MenuData = WithSentryLabel & {
     isBadgeCondensed?: boolean;
 };
 
-type Menu = {
-    sectionStyle: StyleProp<ViewStyle>;
-    sectionTranslationKey: TranslationPaths;
-    items: MenuData[];
-};
+type Menu = {sectionStyle: StyleProp<ViewStyle>; sectionTranslationKey: TranslationPaths; items: MenuData[]};
 
 export type {MenuData};
 
@@ -145,12 +139,8 @@ function InitialSettingsPage({currentUserPersonalDetails}: InitialSettingsPagePr
     const [bankAccountList] = useOnyx(ONYXKEYS.BANK_ACCOUNT_LIST);
     const [fundList] = useOnyx(ONYXKEYS.FUND_LIST);
     const [walletTerms] = useOnyx(ONYXKEYS.WALLET_TERMS);
-    const [loginList] = useOnyx(ONYXKEYS.LOGINS, {
-        selector: expensifyLoginsSelector,
-    });
-    const [hasDeviceManagementErrorValue] = useOnyx(ONYXKEYS.LOGINS, {
-        selector: hasDeviceManagementError,
-    });
+    const [loginList] = useOnyx(ONYXKEYS.LOGINS, {selector: expensifyLoginsSelector});
+    const [hasDeviceManagementErrorValue] = useOnyx(ONYXKEYS.LOGINS, {selector: hasDeviceManagementError});
     const [privatePersonalDetails] = useOnyx(ONYXKEYS.PRIVATE_PERSONAL_DETAILS);
     const [vacationDelegate] = useOnyx(ONYXKEYS.NVP_PRIVATE_VACATION_DELEGATE);
     const allCards = useNonPersonalCardList();
@@ -177,9 +167,7 @@ function InitialSettingsPage({currentUserPersonalDetails}: InitialSettingsPagePr
     const hasActivatedWallet = ([CONST.WALLET.TIER_NAME.GOLD, CONST.WALLET.TIER_NAME.PLATINUM] as string[]).includes(userWallet?.tierName ?? '');
     const hasLockedBankAccount = bankAccountList ? Object.values(bankAccountList).some((bankAccount) => bankAccount.accountData?.state === CONST.BANK_ACCOUNT.STATE.LOCKED) : false;
     const [firstDayFreeTrial] = useOnyx(ONYXKEYS.NVP_FIRST_DAY_FREE_TRIAL);
-    const [isTrackingGPS = false] = useOnyx(ONYXKEYS.GPS_DRAFT_DETAILS, {
-        selector: isTrackingSelector,
-    });
+    const [isTrackingGPS = false] = useOnyx(ONYXKEYS.GPS_DRAFT_DETAILS, {selector: isTrackingSelector});
     const [lastDayFreeTrial] = useOnyx(ONYXKEYS.NVP_LAST_DAY_FREE_TRIAL);
     const [unsharedBankAccount] = useOnyx(ONYXKEYS.UNSHARE_BANK_ACCOUNT);
     const [stashedCredentials] = useOnyx(ONYXKEYS.STASHED_CREDENTIALS);
@@ -201,9 +189,7 @@ function InitialSettingsPage({currentUserPersonalDetails}: InitialSettingsPagePr
         personalCard: {shouldShowRBR: shouldShowRBRForPersonalCard},
     } = useCardFeedErrors();
     const hasPendingCardAction = hasPendingExpensifyCardAction(allCards, privatePersonalDetails);
-    const [isActingAsDelegate] = useOnyx(ONYXKEYS.ACCOUNT, {
-        selector: isActingAsDelegateSelector,
-    });
+    const [isActingAsDelegate] = useOnyx(ONYXKEYS.ACCOUNT, {selector: isActingAsDelegateSelector});
     const hasVirtualCardMissingDetails = hasVirtualExpensifyCardMissingPersonalDetails(allCards, privatePersonalDetails, isActingAsDelegate);
     let walletBrickRoadIndicator;
     if (
@@ -249,81 +235,16 @@ function InitialSettingsPage({currentUserPersonalDetails}: InitialSettingsPagePr
         });
     };
 
-    const showSaveReceiptsModal = (pendingReceiptCount: number) => {
-        return showConfirmModal({
-            title: translate('initialSettingsPage.saveReceiptsConfirmation.title'),
-            prompt: translate('initialSettingsPage.saveReceiptsConfirmation.prompt', {
-                count: pendingReceiptCount,
-            }),
-            confirmText: translate('initialSettingsPage.saveReceiptsConfirmation.confirm'),
-            cancelText: translate('common.cancel'),
-            shouldShowCancelButton: true,
-        });
-    };
-
-    // Combined modal for the offline case: warns about losing offline changes and offers to save the pending receipts, so the user is not shown two back-to-back prompts.
-    const showSaveReceiptsAndSignOutModal = (pendingReceiptCount: number) => {
-        return showConfirmModal({
-            title: translate('initialSettingsPage.saveReceiptsAndSignOutConfirmation.title'),
-            prompt: translate('initialSettingsPage.saveReceiptsAndSignOutConfirmation.prompt', {
-                count: pendingReceiptCount,
-            }),
-            confirmText: translate('initialSettingsPage.saveReceiptsAndSignOutConfirmation.confirm'),
-            cancelText: translate('common.cancel'),
-            shouldShowCancelButton: true,
-            danger: true,
-        });
-    };
-
-    // Save must complete before the forced-signout branch dispatches `Onyx.clear`, which wipes the persisted queue that holds these local file paths.
-    const saveReceipts = async (saveableReceipts: ReturnType<typeof getSaveablePendingReceiptRequests>) => {
-        try {
-            const {savedCount, failedCount} = await saveReceiptsToGallery(saveableReceipts);
-            Log.info('[Receipt] Saved pending receipts to gallery before sign-out', false, {savedCount, failedCount});
-        } catch (error) {
-            Log.alert('[Receipt] Unexpected rejection from saveReceiptsToGallery; sign-out continued', {error});
-        }
-    };
-
     const signOut = async (shouldForceSignout = false) => {
-        // Forced sign-out (expired session, SAML re-auth) must be non-interactive: it must not touch the gallery flow, which can trigger OS permission prompts and delay the redirect.
-        if (shouldForceSignout) {
+        if ((!network.isOffline && !isTrackingGPS) || shouldForceSignout) {
             return signOutAndRedirectToSignIn();
         }
 
-        // `getSaveablePendingReceiptRequests` is platform-split (web returns `[]`) and image-filtered so we do not promise a save the native gallery API can not deliver.
-        const saveableReceipts = getSaveablePendingReceiptRequests();
-        const shouldWarnBeforeSignOut = network.isOffline || isTrackingGPS;
-        // Offline + receipts is the common case; merge the offline warning and the save-receipts prompt into a single modal. GPS keeps its own warning, so it falls through to the two-step path below.
-        const isOfflineReceiptsCase = network.isOffline && !isTrackingGPS && saveableReceipts.length > 0;
-
-        if (!shouldWarnBeforeSignOut && saveableReceipts.length === 0) {
-            return signOutAndRedirectToSignIn();
+        // When offline, warn the user that any actions they took while offline will be lost if they sign out
+        const result = await showSignOutModal();
+        if (result.action !== ModalActions.CONFIRM) {
+            return;
         }
-
-        if (isOfflineReceiptsCase) {
-            const result = await showSaveReceiptsAndSignOutModal(saveableReceipts.length);
-            if (result.action !== ModalActions.CONFIRM) {
-                return;
-            }
-            await saveReceipts(saveableReceipts);
-        } else {
-            if (shouldWarnBeforeSignOut) {
-                const result = await showSignOutModal();
-                if (result.action !== ModalActions.CONFIRM) {
-                    return;
-                }
-            }
-
-            if (saveableReceipts.length > 0) {
-                const result = await showSaveReceiptsModal(saveableReceipts.length);
-                if (result.action !== ModalActions.CONFIRM) {
-                    return;
-                }
-                await saveReceipts(saveableReceipts);
-            }
-        }
-
         if (isTrackingGPS) {
             stopGpsTripNotification();
             stopLocationUpdatesAsync(BACKGROUND_LOCATION_TRACKING_TASK_NAME).catch((error) => console.error('[GPS distance request] Failed to stop location tracking', error));
